\name{GP.un}
\alias{GP.un}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal R-side effects function
}
\description{
An internal function
}
\usage{
GP.un(Z_mat, fixed_effects, control)
}
\arguments{
  \item{Z_mat}{
data frame
}
  \item{fixed_effects}{
formula specifying fixed effects to be included in model
}
\item{control}{
a list
}
}