\name{print}
\alias{print.GPvam}
\alias{print.summary.GPvam}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print
}
\description{
Prints names of elements in GPvam object.
}
\usage{
\method{print}{GPvam}(x, ...)
}
\arguments{
  \item{x}{
object of class GPvam
}
  \item{\dots}{
other arguments to be passed to summary
}
}