\name{rGP.un}
\alias{rGP.un}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal R-side effects function for reduced GP model
}
\description{
An internal function
}
\usage{
rGP.un(Z_mat, fixed_effects, control)
}
\arguments{
  \item{Z_mat}{
data frame
}
  \item{fixed_effects}{
formla specifying fixed effects to be included in model
}
\item{control}{
a list
}
}