\name{GPvam-package}
\alias{GPvam-package}
\docType{package}
\title{
Maximum Likelihood Estimation of the Multiple Membership Mixed Models Used in Value-Added Modeling
}
\description{
An EM algorithm is used to estimate the generalized and variable persistence models (Mariano et al. 2010).
}
\details{
\tabular{ll}{
Package: \tab GPvam\cr
Type: \tab Package\cr
Version: \tab 2.0-0\cr
Date: \tab 2012-07-05\cr
License: \tab GPL-2\cr
}
}
\author{
Andrew Karl, Yan Yang, and Sharon Lohr

Maintainer: Andrew Karl <akarl@asu.edu>
}
\references{
Karl, A., Yang, Y. and Lohr, S. (2012) Efficient Maximum Likelihood Estimation for Multiple Membership Mixed Models Used in Value-Added Modeling. \emph{Submitted.}

Lockwood, J., McCaffrey, D., Mariano, L., Setodji, C. (2007) Bayesian Methods for Scalable Multivariate Value-Added Assesment. \emph{Journal of Educational and Behavioral Statistics} \bold{32}, 125--150. 

Mariano, L., McCaffrey, D. and Lockwood, J. (2010) A Model for Teacher Effects From Longitudinal Data Without Assuming Vertical Scaling. \emph{Journal of Educational and Behavioral Statistics} \bold{35}, 
253--279.

McCaffrey, D. and Lockwood, J. (2011) Missing Data in Value-Added Modeling of Teavher Effects," \emph{Annals of Applied Statistics} \bold{5}, 773--797
}


